mesa: aid users needing libglvnd
This does a simple check during build time for libglvnd and
enables it if present. This prevents users from having to fork
the package to add GL support.

I want to /eventually/ add some kind of "optional" dependency
feature to the package manager so we can be rid of these hacks
once and for all.