nodejs: fix build for LLVM-only systems
This is one of the weirdest shit I had to work with.
If I set export CC=cc CXX=c++ instead of doing it this way, I get linker
error 200 objects in. Also, --partly-static has to be disabled since
it appends -static-libstdc++ and -static-libgcc, which Clang doesn't
support (if both of those libraries are nonexistent).