Small patch for the default config file
Sway default config file includes a line that sets the background
to an image, that is normaly included with sway, using
"output * bg ....". This also requires that swaybg is installed.

Since sway package removes default backgrounds and swaybg
isn't even in the official repos this patch removes references
to nonexistent background images and makes it clear that swaybg
is needed for setting backgrounds.