mold: 0.9.3
See the README file.

---

READMEs like this one are something I want to add to each package.
This solves the lack of a description for each package as well as
the disconnect between Wiki pages and the software itself.

The website can leverage these files to provide full documentation
for each package. This will also make 'kiss help <pkg>' work.